requirements.txt
fastapi
agno
openai
duckduckgo-search
yfinance
sqlalchemy
fastapi[all]
uvicorn
python-dotenv
